requirements.txt
ultralytics
streamlit
torch
Pillow
